trying to get generator to rely on the same methods that recipeformatting does to make my life easier
