--- generator/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2015",
-    "module": "commonjs",
-    "outDir": "js",
-    "sourceMap": true
-  },
-  "include": [
-    "ts/**/*"
-  ]
-}
