+++ ts/GlyphUnit.ts
@@ -1,0 +1,5 @@
+export default interface GlyphUnit {
+    regex: RegExp;
+    replace: string;
+    plaintext: string;
+}
