+++ ts/RegexUnit.ts
@@ -1,0 +1,5 @@
+export default interface RegexUnit {
+    replace: string;
+    regex: RegExp;
+    words: RegExp;
+}
